Refactor frontend structure and remove deprecated components
- Deleted nki_beta2.py and triton.py frontends as they are no longer needed.
- Updated visualization.js to improve error messaging for data loading failures.
- Introduced new frontend registration and runtime patching helpers in __init__.py.
- Created base.py to manage frontend operations, adapters, and patching mechanisms.
- Implemented nki.py and nki_beta2.py frontends with updated adapter functions.
- Established triton.py frontend with comprehensive operation handling and patching capabilities.
