+++ docs/index.html
@@ -1149,7 +1149,7 @@ Consider enabling Buffer Load optimization.
                                                         </div>
                                                         <!-- Sanitizer Client -->
                                                         <div class="client-content" id="client-san">
-                                                            <p class="visualizer-note">The sanitizer symbolically checks tensor memory accesses for out-of-bounds errors and emits rich reports with tensor metadata, call stack, and expression trees, with an optional fake-memory backend to avoid real reads.</p>
+                                                            <p class="visualizer-note">The sanitizer symbolically checks tensor memory accesses for out-of-bounds errors and emits rich reports with tensor metadata, call stack, and expression trees, with optional fake-memory storage to avoid real reads.</p>
                                                             <!-- Sanitizer Sub-Tabs -->
                                                             <div class="features-tabs" style="border: none; padding-bottom: 1rem; justify-content: start;">
                                                                 <button class="tab-btn active" onclick="switchSanitizer(event, 'san-pass')">No Out-of-Bounds</button>
