+++ examples/nki/rmsnorm.py
@@ -87,7 +87,7 @@ def _run_demo():
 
     if triton_viz_enabled:
         print("Executing kernel with NKI interpreter...")
-        traced_kernel = triton_viz.trace(client=Tracer(), backend="nki")(
+        traced_kernel = triton_viz.trace(client=Tracer(), frontend="nki")(
             nki_rmsnorm_kernel
         )
         kernel_instance = traced_kernel[kernel_grid]
