+++ examples/nki/rope.py
@@ -231,7 +231,7 @@ def _run_demo():
         )
 
         print("Executing rotary embedding kernel with NKI interpreter...")
-        traced_kernel = triton_viz.trace(client=Tracer(), backend="nki")(
+        traced_kernel = triton_viz.trace(client=Tracer(), frontend="nki")(
             nki_rope_kernel
         )
         kernel_instance = traced_kernel[kernel_grid]
