+++ examples/nki_beta2/rope.py
@@ -219,7 +219,7 @@ def _run_demo():
     if TRITON_VIZ_ENABLED:
         import triton_viz
 
-        traced_kernel = triton_viz.trace("tracer", backend="nki_beta2")(rope_kernel)
+        traced_kernel = triton_viz.trace("tracer", frontend="nki_beta2")(rope_kernel)
         traced_kernel[kernel_grid](*kernel_args, pre_trace=PRE_TRACE)
         assert np.allclose(expected_q, out_q4d)
         assert np.allclose(expected_k, out_k4d)
