+++ frontend/core/visualization.ts
@@ -644,7 +644,7 @@ async function fetchData(): Promise<void> {
         initializeUIElements();
     } catch (error) {
         console.error('Error fetching data:', error);
-        showControlToast('Failed to load data. Please ensure the backend is running.');
+        showControlToast('Failed to load data. Please ensure the server is running.');
     }
 }
