+++ tests/nki/test_nki_tracer.py
@@ -27,7 +27,7 @@ def div_ceil(n, d):
 def test_tracer_records_masked_load_store():
     triton_viz.clear()
 
-    @triton_viz.trace(client=Tracer(), backend="nki")
+    @triton_viz.trace(client=Tracer(), frontend="nki")
     def add_kernel(x_ptr, y_ptr, out_ptr):
         block_size = 4
         pid = nl.program_id(axis=0)
@@ -73,7 +73,7 @@ def copy_kernel(x_ptr, out_ptr):
 def test_tracer_grid_idx_sampling():
     triton_viz.clear()
 
-    traced = triton_viz.trace(client=Tracer(grid_idx=1), backend="nki")(copy_kernel)
+    traced = triton_viz.trace(client=Tracer(grid_idx=1), frontend="nki")(copy_kernel)
 
     n_elements = 12
     x = NDArray(value=np.arange(n_elements, dtype=np.float32))
@@ -96,7 +96,7 @@ def test_tracer_grid_idx_sampling():
 def test_tracer_records_reduce_sum():
     triton_viz.clear()
 
-    @triton_viz.trace(client=Tracer(), backend="nki")
+    @triton_viz.trace(client=Tracer(), frontend="nki")
     def reduce_sum_kernel(x_ptr, out_ptr):
         block_m = 4
         block_n = 8
@@ -137,7 +137,7 @@ def test_tracer_records_dot():
     TILE_K = 2
     TILE_N = 4
 
-    @triton_viz.trace(client=Tracer(), backend="nki")
+    @triton_viz.trace(client=Tracer(), frontend="nki")
     def matmul_kernel(lhs, rhs, result):
         """NKI matmul_kernel to compute a matrix multiplication operation in a tiled manner
 
@@ -212,7 +212,7 @@ def test_tracer_records_dot():
     kernel_args = (lhs_small, rhs_small, result)
 
     print("Executing matmul_kernel with NKI interpreter...")
-    traced_kernel = triton_viz.trace(client=Tracer(), backend="nki")(matmul_kernel)
+    traced_kernel = triton_viz.trace(client=Tracer(), frontend="nki")(matmul_kernel)
     kernel_instance = traced_kernel[kernel_grid]
     kernel_instance(*kernel_args)
 
@@ -229,7 +229,7 @@ def test_tracer_records_dot():
 def test_tracer_records_dot_transpose_x_kwarg():
     triton_viz.clear()
 
-    @triton_viz.trace(client=Tracer(), backend="nki")
+    @triton_viz.trace(client=Tracer(), frontend="nki")
     def dot_kernel(lhs, rhs, out):
         out[...] = nl.matmul(lhs, rhs, transpose_x=True)
 
@@ -253,7 +253,7 @@ def test_nki_trace_save_load_roundtrip(tmp_path: Path):
     """NKI traces should serialize and reload through the shared .tvz path."""
     triton_viz.clear()
 
-    traced = triton_viz.trace(client=Tracer(), backend="nki")(copy_kernel)
+    traced = triton_viz.trace(client=Tracer(), frontend="nki")(copy_kernel)
 
     n_elements = 6
     x = NDArray(value=np.arange(n_elements, dtype=np.float32))
