+++ triton_viz/core/data.py
@@ -41,7 +41,7 @@ class Store(Op):
     masks: npt.NDArray[np.bool_]
     mem_src: str = "SBUF"
     mem_dst: str = "HBM"
-    backend: str = "nki"
+    frontend: str = "nki"
     bytes: int = 0
     time_idx: int = 0
 
@@ -60,7 +60,7 @@ class Load(Op):
     # buffer: str
     mem_src: str = "HBM"
     mem_dst: str = "SBUF"
-    backend: str = "nki"
+    frontend: str = "nki"
     bytes: int = 0
     time_idx: int = 0
