+++ triton_viz/core/frontend/__init__.py
@@ -1,0 +1,2 @@
+"""DSL frontend registration and runtime patching helpers."""
+
