+++ triton_viz/core/frontend/base.py
@@ -1,0 +1,198 @@
+from collections.abc import Callable
+from contextlib import nullcontext
+from dataclasses import dataclass
+from importlib import import_module
+from typing import Any
+
+from triton_viz.core.data import Op
+
+
+_MISSING = object()
+
+
+class _LangPatchScope:
+    """Tracks patched attributes and mapping items so they can be restored."""
+
+    def __init__(self) -> None:
+        self._changes: list[tuple[str, object, str, object]] = []
+
+    def set_attr(self, obj: object, name: str, value: object) -> None:
+        original = getattr(obj, name, _MISSING)
+        self._changes.append(("attr", obj, name, original))
+        setattr(obj, name, value)
+
+    def set_item(self, mapping: dict[str, Any], key: str, value: object) -> None:
+        original = mapping.get(key, _MISSING)
+        self._changes.append(("item", mapping, key, original))
+        mapping[key] = value
+
+    def restore(self) -> None:
+        while self._changes:
+            kind, obj, name, original = self._changes.pop()
+            if kind == "item":
+                mapping = obj
+                if original is _MISSING:
+                    del mapping[name]
+                else:
+                    mapping[name] = original
+            elif original is _MISSING:
+                delattr(obj, name)
+            else:
+                setattr(obj, name, original)
+
+
+class LoopPatcher:
+    """Manages frontend AST patching for for-loop interception."""
+
+    def __init__(
+        self,
+        transformer_cls: type,
+        visit_for: Callable,
+        method_name: str = "visit_For",
+    ):
+        self._transformer_cls = transformer_cls
+        self._visit_for = visit_for
+        self._method_name = method_name
+        self._orig_visit_for: Callable | None = None
+        self._patched: bool = False
+
+    def patch(self) -> None:
+        if not self._patched:
+            self._orig_visit_for = getattr(
+                self._transformer_cls,
+                self._method_name,
+                None,
+            )
+            setattr(self._transformer_cls, self._method_name, self._visit_for)
+            self._patched = True
+
+    def unpatch(self) -> None:
+        if not self._patched:
+            return
+        if self._orig_visit_for is not None:
+            setattr(self._transformer_cls, self._method_name, self._orig_visit_for)
+        self._patched = False
+
+
+@dataclass
+class AdapterResult:
+    """
+    Standardized callback arguments derived from frontend-specific op signatures.
+    """
+
+    args: tuple[Any, ...]
+    kwargs: dict[str, Any]
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        self.args = args
+        self.kwargs = kwargs
+
+
+def _missing_op_stub(attr: str, namespace: Any) -> Callable:
+    """Create a placeholder for a missing op that raises at runtime."""
+
+    def stub(*args: Any, **kwargs: Any) -> Any:
+        raise AttributeError(
+            f"triton-viz failed to intercept '{attr}': "
+            f"'{type(namespace).__name__}' has no attribute '{attr}'. "
+            "This operation is not supported by your frontend version."
+        )
+
+    return stub
+
+
+def passthrough_adapter(*args: Any, **kwargs: Any) -> AdapterResult:
+    """Return arguments unchanged for clients that expect the original signature."""
+    return AdapterResult(*args, **kwargs)
+
+
+@dataclass
+class Frontend:
+    """
+    Defines one DSL frontend's patchable ops, adapters, and runtime patch hooks.
+    """
+
+    name: str
+    builder: Any
+    original_ops: dict[Any, dict[str, Callable]]
+    adapters: dict[type[Op], Callable[..., AdapterResult]]
+    namespaces: dict[Any, dict[str, type[Op]]]
+
+    @classmethod
+    def from_namespaces(
+        cls,
+        *,
+        name: str,
+        builder: Any,
+        namespaces: dict[Any, dict[str, type[Op]]],
+        adapters: dict[type[Op], Callable[..., AdapterResult]],
+    ) -> "Frontend":
+        original_ops: dict[Any, dict[str, Callable]] = {
+            namespace: {} for namespace in namespaces
+        }
+        for namespace, attrs in namespaces.items():
+            for attr in attrs:
+                if hasattr(namespace, attr):
+                    original_ops[namespace][attr] = getattr(namespace, attr)
+                else:
+                    original_ops[namespace][attr] = _missing_op_stub(attr, namespace)
+
+        for attrs in namespaces.values():
+            for op_type in attrs.values():
+                adapters.setdefault(op_type, passthrough_adapter)
+
+        return cls(
+            name=name,
+            builder=builder,
+            original_ops=original_ops,
+            adapters=adapters,
+            namespaces=namespaces,
+        )
+
+    def maybe_yield_for_multism(self) -> None:
+        return None
+
+    def run_op_overrider(
+        self,
+        op: Callable,
+        op_type: type[Op],
+        callbacks: Any,
+        args: tuple[Any, ...],
+        kwargs: dict[str, Any],
+    ):
+        return callbacks.op_overrider(*args, **kwargs)
+
+    def patch_for_loop(self) -> None:
+        return None
+
+    def unpatch_for_loop(self) -> None:
+        return None
+
+    def patch_lang(self, fn: Callable, client_manager: Any = None) -> _LangPatchScope:
+        return _LangPatchScope()
+
+    def patch_calls(self):
+        return nullcontext()
+
+
+FRONTEND_REGISTRY: dict[str, Frontend] = {}
+LANG_PATCH_SCOPES: dict[str, list[Any]] = {
+    "triton": [],
+    "nki": [],
+    "nki_beta2": [],
+}
+
+
+def register_frontend(frontend: Frontend) -> Frontend:
+    FRONTEND_REGISTRY[frontend.name] = frontend
+    LANG_PATCH_SCOPES.setdefault(frontend.name, [])
+    return frontend
+
+
+def get_frontend(name: str) -> Frontend:
+    if name not in FRONTEND_REGISTRY:
+        import_module(f"triton_viz.core.frontend.{name}")
+    try:
+        return FRONTEND_REGISTRY[name]
+    except KeyError as exc:
+        raise ValueError(f"Unknown frontend: {name}") from exc
