--- triton_viz/frontends/__init__.py
@@ -1,1 +1,0 @@
-from . import nki, nki_beta2, triton  # noqa: F401
